feat: add Deep Wiki modal for enhanced research navigation and visualization
