+++ src/renderer/src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useApp } from '../store/app';
 import { useDocuments } from '../store/documents';
 import { useDeepSearch } from '../store/deepSearch';
@@ -9,12 +9,14 @@ import { ChatPanel } from './ChatPanel';
 import { TableOfContents } from './TableOfContents';
 import { SourcePreviewPopup } from './SourcePreview';
 import { DeepSearchModal } from './research/DeepSearchModal';
+import { DeepWikiModal } from './DeepWikiModal';
 import logoUrl from '../assets/logo.svg';
 
 export function Layout(): JSX.Element {
   const { project, openSettings, closeProject } = useApp();
   const loadFiles = useDocuments((s) => s.loadFiles);
   const openDeepSearch = useDeepSearch((s) => s.open);
+  const [showDeepWiki, setShowDeepWiki] = useState(false);
 
   useEffect(() => {
     loadFiles().catch(console.error);
@@ -29,6 +31,9 @@ export function Layout(): JSX.Element {
           {project && <span className="project-name">· {project.name}</span>}
         </div>
         <div className="titlebar-right">
+          <button type="button" className="titlebar-btn" onClick={() => setShowDeepWiki(true)}>
+            Deep Wiki
+          </button>
           <button type="button" className="titlebar-btn" onClick={openDeepSearch}>
             Deep Search
           </button>
@@ -55,6 +60,7 @@ export function Layout(): JSX.Element {
       </main>
       <SourcePreviewPopup />
       <DeepSearchModal />
+      {showDeepWiki && <DeepWikiModal onClose={() => setShowDeepWiki(false)} />}
     </div>
   );
 }
