+++ src/renderer/src/components/SourcesPanel.tsx
@@ -2,7 +2,6 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import type { SourceMeta } from '@shared/types';
 import { useSourcePreview } from '../store/sourcePreview';
 import { bridge } from '../api/bridge';
-import { WikiGraphModal } from './WikiGraphModal';
 
 type AddMode = null | 'options' | 'paste';
 
@@ -10,7 +9,6 @@ export function SourcesPanel(): JSX.Element {
   const [sources, setSources] = useState<SourceMeta[]>([]);
   const [ingesting, setIngesting] = useState(false);
   const [showAddPopup, setShowAddPopup] = useState(false);
-  const [showGraph, setShowGraph] = useState(false);
   const openPreview = useSourcePreview((s) => s.open);
 
   const loadSources = useCallback(() => {
@@ -32,17 +30,6 @@ export function SourcesPanel(): JSX.Element {
     <div className="sources-panel">
       <div className="sources-panel-header">
         <h2>Sources</h2>
-        {sources.length > 0 && (
-          <button
-            type="button"
-            className="wiki-graph-trigger"
-            onClick={() => setShowGraph(true)}
-            title="Show research wiki graph"
-            aria-label="Show research wiki graph"
-          >
-            &#x25CE;
-          </button>
-        )}
       </div>
 
       {sources.length > 0 && (
@@ -97,8 +84,6 @@ export function SourcesPanel(): JSX.Element {
           onDone={() => setIngesting(false)}
         />
       )}
-
-      {showGraph && <WikiGraphModal onClose={() => setShowGraph(false)} />}
     </div>
   );
 }
