+++ src/renderer/src/styles.css
@@ -2488,65 +2488,10 @@ img.login-logo {
 .sources-panel-header h2 {
   margin: 0;
 }
-.wiki-graph-trigger {
-  background: none;
-  border: 1px solid var(--border);
-  color: var(--fg-muted);
-  width: 26px;
-  height: 26px;
-  border-radius: 50%;
-  cursor: pointer;
-  font-size: 14px;
-  line-height: 1;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  transition: border-color 0.15s, color 0.15s, background 0.15s;
-}
-.wiki-graph-trigger:hover {
-  border-color: var(--accent);
-  color: var(--accent);
-  background: rgba(var(--accent-rgb, 100, 149, 237), 0.08);
-}
-
-/* Wiki graph modal */
-.wiki-graph-overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.58);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 120;
-  backdrop-filter: blur(2px);
-}
-.wiki-graph-modal {
-  background: var(--bg-panel);
-  border: 1px solid var(--border);
-  border-radius: 12px;
-  padding: 20px 22px 18px;
-  width: min(820px, 92vw);
-  max-height: 88vh;
-  display: flex;
-  flex-direction: column;
-  box-shadow: 0 14px 40px rgba(0, 0, 0, 0.45);
-}
-.wiki-graph-header {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 12px;
-  margin-bottom: 10px;
-}
-.wiki-graph-header h3 {
-  margin: 0;
-  font-size: 16px;
-  font-weight: 600;
-}
-.wiki-graph-sub {
+/* Deep Wiki graph canvas — mounted inside the ds-modal shell */
+.ds-modal-stats {
   font-size: 11px;
-  color: var(--fg-muted);
-  margin-top: 2px;
+  margin-bottom: 10px;
 }
 .wiki-graph-canvas-wrap {
   flex: 1;
@@ -2635,20 +2580,6 @@ img.login-logo {
   font-size: 12px;
   color: var(--fg-muted);
   line-height: 1.4;
-}
-.wiki-graph-footer {
-  margin-top: 10px;
-  font-size: 11px;
-  color: var(--fg-muted);
-  text-align: center;
-}
-.wiki-graph-footer code {
-  font-family: var(--font-mono, monospace);
-  font-size: 11px;
-  background: var(--bg);
-  padding: 1px 5px;
-  border-radius: 3px;
-  border: 1px solid var(--border);
 }
 
 /* ---- Bug report modal ---- */
